LayoutTests/imported/w3c/web-platform-tests/2dcontext/text-styles/2d.text.draw.space.collapse.other-expected.txt

2d.text.draw.space.collapse.other
Space characters are converted to U+0020, and collapsed (per CSS)
Actual output:
Expected output:


FAIL Space characters are converted to U+0020, and collapsed (per CSS) assert_approx_equals: Red channel of the pixel at (25, 25) expected 0 +/- 2 but got 255
